Move parse_boolean to core.helpers
